Move file handling to separate file
