+++ lib/server/home.js
@@ -1,33 +1,8 @@
-import {promises as fs} from 'fs'
+import {File, serveStaticFile} from './file.js'
 import {readLatestSites} from '../db/stats.js'
 import {createHome} from '../page/home.js'
 import {createOptions} from '../page/options.js'
 
-const filesByPath = new Map()
-
-
-class File {
-	constructor(path, mime) {
-		this.path = path
-		this.mime = mime
-		this.contents = null
-		this.lastModified = 0
-	}
-
-	async readContents() {
-		if (!this.contents) {
-			this.contents = await fs.readFile(this.path)
-			this.lastModified = (await fs.stat(this.path)).mtimeMs
-		}
-		const since = (await fs.stat(this.path)).mtimeMs
-		if (since > this.lastModified) {
-			console.log(`Reloading ${this.path}`)
-			this.contents = await fs.readFile(this.path)
-			this.lastModified = since
-		}
-		return this.contents
-	}
-}
 
 export default async function setup(app) {
 	app.get('/', serveHome)
@@ -45,20 +20,10 @@ async function serveFile(path, mime) {
 	}
 }
 
-async function serveStaticFile(file, request, reply) {
-	const contents = await file.readContents()
-	reply.type(file.mime)
-	return contents
-}
-
 async function serveSvgLogo(request, reply) {
 	const path = `public/img/${request.params.file}.svg`
 	const mime = 'image/svg+xml'
-	if (!filesByPath.has(path)) {
-		filesByPath.set(path, new File(path, mime))
-	}
-	const file = filesByPath.get(path)
-	return serveStaticFile(file, request, reply)
+	return serveStaticFile(path, mime, request, reply)
 }
 
 /**
